requirements.txt
absl-py
numpy>=2.2.6
jupyter
openai>=1.3.1
dotenv>=0.9.9
